requirements.txt
pyopengl
cython
yacs
opencv-python
PyYAML
cython-bbox
scipy
progress
motmetrics
matplotlib
lap
openpyxl
Pillow
tensorboardX
fvcore
numba
torch
torchvision
motmetrics
visdom